Restructure file system and rename files
